+++ src/workbench/extensions/manager/utils/conflictUtils.ts
