+++ src/workbench/extensions/manager/utils/systemCompatibility.ts
