+++ tests-ui/tests/utils/versionUtil.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it, vi } from 'vitest'
 import {
   checkVersionCompatibility,
   getFrontendVersion
-} from '@/utils/versionUtil'
+} from '@/workbench/extensions/manager/utils/versionUtil'
 
 // Mock config module
 vi.mock('@/config', () => ({
@@ -287,7 +287,9 @@ describe('versionUtil', () => {
       vi.resetModules()
 
       // Import fresh module
-      const versionUtil = await import('@/utils/versionUtil')
+      const versionUtil = await import(
+        '@/workbench/extensions/manager/utils/versionUtil'
+      )
 
       const version = versionUtil.getFrontendVersion()
       expect(version).toBe('2.0.0')
@@ -320,7 +322,9 @@ describe('versionUtil', () => {
       vi.resetModules()
 
       // Import fresh module
-      const versionUtil = await import('@/utils/versionUtil')
+      const versionUtil = await import(
+        '@/workbench/extensions/manager/utils/versionUtil'
+      )
 
       const version = versionUtil.getFrontendVersion()
       expect(version).toBeUndefined()
